refactor: 타입/상수 정의 코드 올바른 디렉토리로 이동
