+++ service/malls.ts
@@ -1,34 +1,5 @@
+import { Mall, MallPreview, MallRankingPreview } from '@/types/malls';
 import { customFetch, serverFetch } from '@/utils/customFetch';
-
-export type MallRankingPreview = {
-  id: number;
-  name: string;
-  image: string;
-  isFavorite: boolean;
-};
-
-export type MallPreview = {
-  id: number;
-  name: string;
-  image: string;
-  description: string;
-  view: number;
-  isFavorite: boolean;
-  products: {
-    productId: number;
-    productImage: string;
-    productName: string;
-  }[];
-};
-
-export type Mall = {
-  id: number;
-  name: string;
-  url: string;
-  image: string;
-  description: string;
-  isFavorite: boolean;
-};
 
 export async function getRankedMallPreview(): Promise<MallRankingPreview[]> {
   const response = await customFetch('/malls/rank/preview');
