+++ service/mysize.ts
@@ -1,24 +1,5 @@
+import { MySize, smartAnalysisRequestBody } from '@/types/mysize';
 import { customFetch } from '@/utils/customFetch';
-
-export type MySize = {
-  height: number | null;
-  weight: number | null;
-  arm: number | null;
-  leg: number | null;
-  shoulder: number | null;
-  waist: number | null;
-  chest: number | null;
-  thigh: number | null;
-  hip: number | null;
-};
-
-export type smartAnalysisRequestBody = {
-  front: string;
-  side: string;
-  height: number;
-  weight: number;
-  sex: string;
-};
 
 export async function getMySize(): Promise<MySize> {
   const response = await customFetch('/users/mysize');
