+++ service/search.ts
@@ -1,21 +1,5 @@
+import { Keywords } from '@/types/search';
 import { serverFetch } from '@/utils/customFetch';
-
-export type Products = {
-  id: number;
-  name: string;
-  image: string;
-};
-
-export type Malls = {
-  id: number;
-  name: string;
-  image: string;
-};
-
-export type Keywords = {
-  products: Products[];
-  malls: Malls[];
-};
 
 export async function getSearchKeywords(keyword: string): Promise<Keywords> {
   const response = await serverFetch(`/search/${keyword}`);
