+++ types/menu.ts
@@ -1,0 +1,4 @@
+export type Menu = {
+  name: string;
+  path: string;
+};
