+++ types/search.ts
@@ -1,0 +1,16 @@
+export type Products = {
+  id: number;
+  name: string;
+  image: string;
+};
+
+export type Malls = {
+  id: number;
+  name: string;
+  image: string;
+};
+
+export type Keywords = {
+  products: Products[];
+  malls: Malls[];
+};
