+++ types/user.ts
@@ -1,3 +1,13 @@
+export type User = {
+  username: string;
+  password: string;
+  email: string;
+  gender: string;
+  year: number;
+  month: number;
+  day: number;
+};
+
 export type MyPage = {
   username: string;
   email: string;
